requirements.txt
Django==4.2.1
djangorestframework==3.14.0
drf-spectacular==0.27.0
factory-boy==3.3.0
django-environ==0.11.2
psycopg2-binary==2.9.9

# Code quality
# ------------------------------------------------------------------------------
flake8==6.1.0  # https://github.com/PyCQA/flake8
flake8-isort==6.1.1  # https://github.com/gforcada/flake8-isort
coverage==7.3.2  # https://github.com/nedbat/coveragepy
black==23.11.0  # https://github.com/psf/black
djlint==1.34.0  # https://github.com/Riverside-Healthcare/djLint
pylint-django==2.5.5  # https://github.com/PyCQA/pylint-django
pre-commit==3.5.0
